Added a link to the official list of drivers
I thought it would be helpful to add a link to a list of other drivers that are available. I thought it best to link to the official list, but feel free to suggest a different / better source. I also edited out a couple comma splices and attempted to use less opinionated language while maintaining the same emphasis on the best practices described.